requirements.txt
tensorflow==2.5.0
streamlit==0.82.0
opencv-python==4.1.2.30
numpy==1.19.5
Pillow==7.1.2